pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comfy-load-last-image"
version = "1.0.0"
description = "A ComfyUI node designed to load the most recent image in a folder"
readme = "README.md"
authors = [
    {name = "Ben Staniford"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["comfyui", "image", "node", "custom"]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "Pillow>=8.0.0",
    "numpy>=1.21.0",
]

[project.urls]
Homepage = "https://github.com/your-username/comfy-load-last-image"
Repository = "https://github.com/your-username/comfy-load-last-image"
Issues = "https://github.com/your-username/comfy-load-last-image/issues"
